+++ src/libs6rc/s6rc_read_uint.c
@@ -1,5 +1,6 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <errno.h>
 #include <skalibs/uint.h>
 #include <skalibs/bytestr.h>
@@ -8,7 +9,7 @@
 int s6rc_read_uint (char const *file, unsigned int *u)
 {
   char buf[UINT_FMT + 1] ;
-  register int r = openreadnclose(file, buf, UINT_FMT) ;
+  register ssize_t r = openreadnclose(file, buf, UINT_FMT) ;
   if (r < 0) return (errno == ENOENT) ? 0 : -1 ;
   buf[byte_chr(buf, r, '\n')] = 0 ;
   if (!uint0_scan(buf, u)) return (errno = EINVAL, -1) ;
